requirements.txt
Pillow==9.1.0
requests==2.27.1
streamlit==1.9.0
streamlit_lottie==0.0.3
